Combine shape indicators and other overlays.
